UDP client and server
This program is a simple socket layer and a reliable transfer service on
top of the UDP/IP protocol which is a service that
guarantees the arrival of datagrams in the correct order on top of the
UDP/IP protocol, along with congestion control.this
program uses techniques to guarantee the arrival of data correctly like
selective repeat , stop and wait and go back n and the
results shown that selective repeat technique is the best of them.
This assignment discusses Reliable UDP (RUDP).
RUDP is a simple packet based transport protocol. RUDP is based on RFCs
1151 and 908 - Reliable Data Protocol. RUDP
is layered on the UDP/IP Protocols and provides reliable in-order
delivery (up to a maximum number of retransmissions) for
virtual connections. RUDP has a very flexible design that would make it
suitable for a variety of transport uses. One such use
would be to transport telecommunication signaling protocols.